file upload

+++ src/modules/users/services/utils/logoutUser.ts
@@ -11,19 +11,25 @@ export const logoutUser = async (
   res,
 ) => {
   try {
-    console.log(refreshToken);
-    const deleted = await tokenModel.destroy({
-      where: { token: refreshToken },
-    });
-
-    if (!deleted) {
+    if (refreshToken) {
+      const deleted = await tokenModel.destroy({
+        where: { token: refreshToken },
+      });
+      if (!deleted) {
+        throw new UnauthorizedException('Пользователь не авторизован');
+      }
+      res.clearCookie('accessToken', {
+        httpOnly: true,
+        secure: false,
+        sameSite: 'lax',
+      });
+      return res
+        .status(HttpStatus.OK)
+        .json({ message: 'Выход из аккаунта выполнен успешно' });
+    } else {
       throw new UnauthorizedException('Пользователь не авторизован');
     }
-    res.clearCookie('accessToken');
   } catch (err) {
-    if (err instanceof UnauthorizedException) {
-      throw err;
-    }
     console.error('Ошибка при выходе пользователя', err);
     throw new HttpException(
       'Ошибка при выходе пользователя',
